chore(robots): rename g1 kinematics file
