+++ src/lerobot/teleoperators/unitree_g1/exo_ik.py
@@ -25,8 +25,8 @@ from dataclasses import dataclass
 
 import numpy as np
 
+from lerobot.robots.unitree_g1.g1_kinematics import G1_29_ArmIK
 from lerobot.robots.unitree_g1.g1_utils import G1_29_JointArmIndex
-from lerobot.robots.unitree_g1.robot_kinematic_processor import G1_29_ArmIK
 
 from .exo_calib import JOINTS
